requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.6
django-admin-flexselect==0.4.1
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-dynamic-preferences==1.16.0
django-filter==23.5
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-multiupload==0.6.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
kombu==5.3.5
packaging==23.2
persisting-theory==1.0
phonenumbers==8.13.28
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
